moved features from branchname to have all features on the same level in the strkture. Removed unnecesary mapping now that our data will look like expected
